backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9
pydantic[email]==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
python-dotenv==1.0.1
aiofiles==23.2.1
slowapi==0.1.9
email-validator==2.3.0